+++ backend/src/main/java/com/kotonekanno/movielog/security/CustomUserDetailsService.java
@@ -1,0 +1,33 @@
+package com.kotonekanno.movielog.security;
+
+import com.kotonekanno.movielog.entity.User;
+import com.kotonekanno.movielog.repository.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+public class CustomUserDetailsService implements UserDetailsService {
+
+  @Autowired
+  private UserRepository userRepository;
+
+  @Override
+  public UserDetails loadUserByUsername(String email) {
+    User user = userRepository.findByEmail(email)
+        .orElseThrow(() -> new UsernameNotFoundException("User not found"));
+
+    if (!user.getIsActive()) {
+      throw new UsernameNotFoundException("This account is not activated");
+    }
+
+    return org.springframework.security.core.userdetails.User
+        .withUsername(user.getEmail())
+        .password(user.getPasswordHash())
+        .roles("USER")
+        .disabled(user.getDeletedAt() != null)
+        .build();
+  }
+}
